% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{spark.bin}
\alias{spark.bin}
\title{Column Binner for Spark}
\usage{
spark.bin(df, binners)
}
\arguments{
\item{df}{(dataframe, tibble, tbl_spark) table to be binned}

\item{binners}{(list) list containing:
\code{input_col (character)}: colname to be binned, must have numeric values
\code{output_col (character)}: column name to be generated as a result of binning
\code{ splits (named numeric)}: named numeric vector containing splits and labels for them}
}
\description{
Bins numeric columns of a given table and creates a new column containing categorical bins
}
